+++ test/blackbox-tests/test-cases/build-number/run.t
@@ -1,0 +1,10 @@
+  $ $JBUILDER build --root . -j 1 --display quiet @x
+  Build number: 0
+  $ $JBUILDER build --root . -j 1 --display quiet @x
+  Build number: 1
+  $ $JBUILDER build --root . -j 1 --display quiet @x
+  Build number: 2
+  $ $JBUILDER build --root . -j 1 --display quiet @x
+  Build number: 3
+  $ $JBUILDER build --root . -j 1 --display quiet @x
+  Build number: 4